streamlit/requirements.txt
pandas
streamlit
psycogpg2
sqlalchemy
python-dotenv